// src/schemas/interfaces/event.d.ts
export interface IEventSelectableFields {
  name?: number;

  vendor?: number;

  state?: number;

  description?: number;

  category?: number;

  country?: number;

  loc?: number;

  location?: number;

  city?: number;

  date?: number;

  capacity?: number;

  suspended?: number;

  photos?: number;

  price?: number;

  isActive?: number;

  isDeleted?: number;

  deletedAt?: number;

  bannerImages?: number;
}